MCTSF_NEW_Clean/requirements.txt
linear_attention_transformer==0.19.1
matplotlib==3.7.0
numpy==1.26.4
pandas==1.5.3
PyYAML==6.0.2
Requests==2.32.4
scikit_learn==1.2.2
seaborn==0.13.2
torch==2.5.1
tqdm==4.65.0
transformers==4.51.3
faiss-cpu